+++ nodeconductor/users/serializers.py
@@ -30,6 +30,7 @@ class InvitationSerializer(serializers.HyperlinkedModelSerializer):
         required=False,
         allow_null=True
     )
+    project_name = serializers.ReadOnlyField(source='project.name')
     customer = serializers.HyperlinkedRelatedField(
         source='customer_role.customer',
         view_name='customer-detail',
@@ -49,12 +50,17 @@ class InvitationSerializer(serializers.HyperlinkedModelSerializer):
         required=False,
         allow_null=True
     )
+    customer_name = serializers.ReadOnlyField(source='customer.name')
+
+    expires = serializers.DateTimeField(source='get_expiration_time', read_only=True)
 
     class Meta(object):
         model = models.Invitation
-        fields = ('url', 'uuid', 'state', 'link_template', 'email',
-                  'project', 'project_role', 'customer', 'customer_role', 'modified', 'created')
-        read_only_fields = ('url', 'uuid', 'state', 'modified', 'created')
+        fields = ('url', 'uuid', 'link_template', 'email', 'civil_number',
+                  'project', 'project_role', 'project_name',
+                  'customer', 'customer_role', 'customer_name',
+                  'state', 'error_message', 'created', 'expires')
+        read_only_fields = ('url', 'uuid', 'state', 'error_message', 'created', 'expires')
         view_name = 'user-invitation-detail'
         extra_kwargs = {
             'url': {'lookup_field': 'uuid'},
@@ -71,32 +77,36 @@ class InvitationSerializer(serializers.HyperlinkedModelSerializer):
         if '{uuid}' not in link_template:
             raise serializers.ValidationError({'link_template': "Link template must include '{uuid}' parameter."})
 
-        customer_role = attrs.get('customer_role', {})
         project_role = attrs.get('project_role', {})
         project = project_role.get('project')
+        project_role_type = project_role.get('role_type')
+
+        customer_role = attrs.get('customer_role', {})
         customer = customer_role.get('customer')
+        customer_role_type = customer_role.get('role_type')
+
         if customer and project:
             raise serializers.ValidationError('Cannot create invitation to project and customer simultaneously.')
         elif not (customer or project):
             raise serializers.ValidationError('Customer or project must be provided.')
-        elif customer and customer_role.get('role_type') is None:
-            raise serializers.ValidationError({'customer_role': 'Customer role must be provided.'})
-        elif project and project_role.get('role_type') is None:
-            raise serializers.ValidationError({'project_role': 'Project role must be provided.'})
+        elif (customer and customer_role_type is None) or (customer_role_type is not None and not customer):
+            raise serializers.ValidationError({'customer_role': 'Customer and its role must be provided.'})
+        elif (project and project_role_type is None) or (project_role_type is not None and not project):
+            raise serializers.ValidationError({'project_role': 'Project and its role must be provided.'})
 
         return attrs
 
     def create(self, validated_data):
         project_role_data = validated_data.pop('project_role', {})
+        project = project_role_data.get('project')
         customer_role_data = validated_data.pop('customer_role', {})
+        customer = customer_role_data.get('customer')
 
-        if project_role_data:
-            project = project_role_data['project']
+        if project:
             project_role = project.roles.get(role_type=project_role_data['role_type'])
             validated_data['project_role'] = project_role
             validated_data['customer'] = project.customer
-        elif customer_role_data:
-            customer = customer_role_data['customer']
+        elif customer:
             customer_role = customer.roles.get(role_type=customer_role_data['role_type'])
             validated_data['customer_role'] = customer_role
             validated_data['customer'] = customer